features/cpu: stop treating CPUID leaf numbers and results as signed
UndefinedBehaviorSanitizer's implicit-conversion check fires twice in
cpu_features_probe() and cpu_features_get_model_name():

  features_cpu.c:940:14: implicit conversion from type 'unsigned int'
  of value 2147483648 to type 'int' changed the value to -2147483648

x86_cpuid() took its leaf number as int, so every extended-leaf call
site -- 0x80000000, 0x80000001, 0x80000002+i -- passed a value with
the top bit set through a signed parameter. Widen it to uint32_t.

That in turn exposed the reverse conversion: flags[] is int32_t but
holds raw register contents, and leaf 0x80000000 legitimately returns
a maximum-leaf value above INT32_MAX, so assigning flags[0] to the
unsigned max_flag converted implicitly. Spell those two out with a
cast rather than changing the array type, which would break source
compatibility for anything outside the tree calling x86_cpuid().

Neither was a live defect -- the values round-trip on every ABI we
build for -- but they made the file impossible to run under
-fsanitize=implicit-conversion, which matters now that the PCLMULQDQ
probe lives here.

Predates the crc32 work; found while sanitizing that series.

Clean afterwards under gcc ASan+LeakSan+UBSan, clang MSan, clang
TSan, clang UBSan+integer+implicit-conversion+unsigned-integer-overflow,
and valgrind memcheck, exercising cpu_features_get(),
cpu_features_get_core_amount() and cpu_features_get_model_name() from
eight threads. Feature detection unchanged: same bitmask, PCLMULQDQ
still reported.

+++ libretro-common/features/features_cpu.c
@@ -285,7 +285,7 @@ retro_time_t cpu_features_get_time_usec(void)
 
 #if defined(CPU_X86) && !defined(__MACH__)
 #include <limits.h>
-void x86_cpuid(int func, int32_t flags[4])
+void x86_cpuid(uint32_t func, int32_t flags[4])
 {
    /* On Android, we compile RetroArch with PIC, and we
     * are not allowed to clobber the ebx register. */
@@ -742,7 +742,10 @@ static uint64_t cpu_features_probe(void)
          && flags[2] == VENDOR_INTEL_c
          && flags[3] == VENDOR_INTEL_d);
 
-   max_flag = flags[0];
+   /* CPUID register contents are unsigned; flags[] is int32_t, so the
+    * widening is spelled out rather than left implicit -- leaf
+    * 0x80000000 legitimately returns values with the top bit set. */
+   max_flag = (uint32_t)flags[0];
    /* Does CPUID not support func = 1? (unlikely ...) */
    if (max_flag < 1) 
       return 0;
@@ -839,7 +842,7 @@ static uint64_t cpu_features_probe(void)
    }
 
    x86_cpuid(0x80000000, flags);
-   max_flag = flags[0];
+   max_flag = (uint32_t)flags[0];
    if (max_flag >= 0x80000001u)
    {
       x86_cpuid(0x80000001, flags);
